Change related: 1. Refine comments to remove the unrecognized code. 2. Add code to avoid potential data type change. 3. In inf file, add module type to the library class list.

git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk/edk2@12284 6f19259b-4bc3-4df7-8a09-765794883524

+++ SecurityPkg/Tcg/TcgSmm/TcgSmm.c
@@ -374,7 +374,7 @@ PublishAcpiTable (
   ASSERT_EFI_ERROR (Status);
 
   ASSERT (Table->OemTableId == SIGNATURE_64 ('T', 'c', 'g', 'T', 'a', 'b', 'l', 'e'));
-  mTcgNvs = AssignOpRegion (Table, SIGNATURE_32 ('T', 'N', 'V', 'S'), sizeof (TCG_NVS));
+  mTcgNvs = AssignOpRegion (Table, SIGNATURE_32 ('T', 'N', 'V', 'S'), (UINT16) sizeof (TCG_NVS));
   ASSERT (mTcgNvs != NULL);
 
   //
